backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
Pillow
sentence-transformers
faiss-cpu
langchain
langchain-community
numpy
scikit-learn
